+++ library/src/methods/fallback/index.ts
@@ -1,0 +1,2 @@
+export { fallback } from './fallback.ts';
+export { fallbackAsync } from './fallbackAsync.ts';