+++ library/src/methods/fallback/types.ts
@@ -1,0 +1,9 @@
+import type { Issues } from '../../error/index.ts';
+
+/**
+ * Fallback info type.
+ */
+export type FallbackInfo = {
+  input: unknown;
+  issues: Issues;
+};