+++ library/src/methods/index.ts
@@ -1,5 +1,6 @@
 export * from './brand/index.ts';
 export * from './coerce/index.ts';
+export * from './fallback/index.ts';
 export * from './is/index.ts';
 export * from './keyof/index.ts';
 export * from './merge/index.ts';